pyproject.toml
[project]
name = "value-investment-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alpha-vantage>=3.0.0",
    "edgartools>=5.8.3",
    "pandas>=2.3.3",
    "requests>=2.32.5",
]
